+++ app/services/payment_providers/stripe/webhooks/customer_updated_service.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module PaymentProviders
-  module Webhooks
-    module Stripe
+  module Stripe
+    module Webhooks
       class CustomerUpdatedService < BaseService
         def call
           return handle_missing_customer unless stripe_customer
