+++ app/services/payment_providers/stripe/webhooks/setup_intent_succeeded_service.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module PaymentProviders
-  module Webhooks
     module Stripe
+      module Webhooks
       class SetupIntentSucceededService < BaseService
         include Customers::PaymentProviderFinder
