+++ spec/services/payment_providers/adyen/webhooks/chargeback_service_spec.rb
@@ -2,14 +2,14 @@
 
 require 'rails_helper'
 
-RSpec.describe PaymentProviders::Webhooks::Stripe::ChargeDisputeClosedService, type: :service do
+RSpec.describe PaymentProviders::Adyen::Webhooks::ChargebackService, type: :service do
   subject(:service) { described_class.new(organization_id:, event_json:) }
 
   let(:organization_id) { organization.id }
   let(:organization) { create(:organization) }
   let(:membership) { create(:membership, organization:) }
   let(:customer) { create(:customer, organization:) }
-  let(:payment) { create(:payment, payable: invoice, provider_payment_id: 'pi_3OzgpDH4tiDZlIUa0Ezzggtg') }
+  let(:payment) { create(:payment, payable: invoice, provider_payment_id: '9915555555555555') }
   let(:lose_dispute_service) { Invoices::LoseDisputeService.new(invoice:) }
   let(:invoice) { create(:invoice, customer:, organization:, status:, payment_status: 'succeeded') }
 
@@ -18,7 +18,7 @@ RSpec.describe PaymentProviders::Webhooks::Stripe::ChargeDisputeClosedService, t
 
     context 'when dispute is lost' do
       let(:event_json) do
-        path = Rails.root.join('spec/fixtures/stripe/charge_dispute_lost_event.json')
+        path = Rails.root.join('spec/fixtures/adyen/chargeback_lost_event.json')
         File.read(path)
       end
 
@@ -54,7 +54,7 @@ RSpec.describe PaymentProviders::Webhooks::Stripe::ChargeDisputeClosedService, t
           end.to have_enqueued_job(SendWebhookJob).with(
             'invoice.payment_dispute_lost',
             payment.payable,
-            provider_error: 'fraudulent'
+            provider_error: 'Merchandise/Services Not Received'
           )
         end
       end
@@ -62,7 +62,7 @@ RSpec.describe PaymentProviders::Webhooks::Stripe::ChargeDisputeClosedService, t
 
     context 'when dispute is won' do
       let(:event_json) do
-        path = Rails.root.join('spec/fixtures/stripe/charge_dispute_won_event.json')
+        path = Rails.root.join('spec/fixtures/adyen/chargeback_won_event.json')
         File.read(path)
       end
